Add examples of multiple sets of radios/checkboxes
We have an open issue to provide examples of grouping radios and checkboxes under headings (https://github.com/alphagov/govuk-frontend/issues/1079).

Whilst we don’t provide any guidance on this at the minute, we’ve previously suggested that users can achieve this by making multiple calls to the `govukCheckboxes` macro and grouping them together inside a single call to `govukFieldset`. However, this means that each ‘set’ of checkboxes is inside its own module, which can cause issues with things like conditional reveals.

Add an example of this to the review app, so that we can test that everything works as expected.