Always initialise modules on radios
This matches a change previously made to the checkboxes component in 7e746b4670e10349a3e678f4bbd6e7d85e8b771d.

If you use two separate calls to the macro with the same name, but one of them does not contain any radios with conditional reaveals, then checking a radio in that list would not hide a conditional reveal in the other list, as the module has not been initialised so no `eventListener` has been set up.

Always initialise the javascript for every set of radios which solves this issue.